moved component-status, component-status-resolver and component-tooltip out of sidebar and added to bit
