+++ components/component-status-resolver/index.ts
