+++ components/component-status/index.ts
@@ -1,0 +1,1 @@
+export { ComponentStatus } from './component-status';
