+++ components/component-tooltip/index.ts
