+++ components/side-bar/index.ts
@@ -1,3 +1,2 @@
 export { ComponentTree } from './component-tree';
 export { Collapser } from './sidebar-collapser';
-export { ComponentStatusResolver } from './component-tree/component-status-resolver';
